requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pytest
httpx
PyYAML
pydantic
